some fixes, but also doesn't work
