+++ GameMarketplace/Market/views/__init__.py
@@ -1,1 +1,1 @@
-from . import auth, market
+from . import auth, market, cart
