+++ GameMarketplace/cart/cart.py
@@ -2,61 +2,75 @@ from decimal import Decimal
 from django.conf import settings
 from Market.models import Product
 
-class Сart(object):
+
+class Cart(object):
 
     def __init__(self, request):
-        
+        """
+        Инициализация корзины
+        """
         self.session = request.session
         cart = self.session.get(settings.CART_SESSION_ID)
         if not cart:
+            # сохраняем ПУСТУЮ корзину в сессии
             cart = self.session[settings.CART_SESSION_ID] = {}
         self.cart = cart
 
     def __iter__(self):
-
+        """
+        Перебираем товары в корзине и получаем товары из базы данных.
+        """
         product_ids = self.cart.keys()
+        # получаем товары и добавляем их в корзину
         products = Product.objects.filter(id__in=product_ids)
 
         cart = self.cart.copy()
-        for product in product:
+        for product in products:
             cart[str(product.id)]['product'] = product
 
         for item in cart.values():
             item['price'] = Decimal(item['price'])
             item['total_price'] = item['price'] * item['quantity']
             yield item
-
+    
     def __len__(self):
+        """
+        Считаем сколько товаров в корзине
+        """
+        return sum(item['quantity'] for item in self.cart.values())
 
-        return sum(item['quality'] for item in self.cart.values())
-
-    def add(self, product, quantity = 1, update_quantity=False):
-
+    def add(self, product, quantity=1, update_quantity=False):
+        """
+        Добавляем товар в корзину или обновляем его количество.
+        """
         product_id = str(product.id)
         if product_id not in self.cart:
             self.cart[product_id] = {'quantity': 0,
-                                    'price': str(product.price)}
+                                      'price': str(product.price)}
         if update_quantity:
             self.cart[product_id]['quantity'] = quantity
         else:
             self.cart[product_id]['quantity'] += quantity
         self.save()
-    
+
     def save(self):
+        # сохраняем товар
         self.session.modified = True
 
     def remove(self, product):
-
+        """
+        Удаляем товар
+        """
         product_id = str(product.id)
         if product_id in self.cart:
             del self.cart[product_id]
             self.save()
-    
+
     def get_total_price(self):
-        
-        return sum(Decimal(item['price']) * item['quanity'] for item in self.cart.values())
+        # получаем общую стоимость
+        return sum(Decimal(item['price']) * item['quantity'] for item in self.cart.values())
 
     def clear(self):
-        
-        del self.session[settings.CART_SESSIN_ID]
+        # очищаем корзину в сессии
+        del self.session[settings.CART_SESSION_ID]
         self.save()
