+++ GameMarketplace/config/urls.py
@@ -19,4 +19,5 @@ from django.urls import path, include
 urlpatterns = [
     path('admin/', admin.site.urls),
     path('', include('Market.urls')),
+    path('cart/', include('cart.urls', namespace='cart')),
 ]
