+++ GameMarketplace/templates/Market/catalog.html
@@ -26,13 +26,8 @@
             {% for product in products %}
             <tr>
                 
-                <td class="product_name"><a href="{% url 'product_info' pk=product.pk %}">{{ product.name }}</td>
+                <td class="product_name"><a href="{% url 'market:product_info' pk=product.pk %}">{{ product.name }}</td>
                 <td>{{ product.price }}</td>
-                    <form action="{% url 'cart:cart_add' product.id %}" method="post">
-                        {{ cart_product_form }}
-                        {% csrf_token %}
-                        <input type="sumbit" value="В коризину">
-                    </form>
                 <td>{{ product.category }}</td>
                 </a>
 
@@ -40,9 +35,11 @@
                     <img src="{{ MEDIA_URL }}{{ product.img }}" width="100" height="100">
                 </td>
                 <td>
-                    <a href="">
-                        <div class="ui red button">Добавить в корзину</div>
-                    </a>
+                    <form action="{% url 'cart:cart_add' product.id %}" method="post">
+                        {{ cart_product_form }}
+                        {% csrf_token %}
+                        <button type="sumbit" value="В коризину" class="ui red button">В корзину
+                    </form>
                 </td>
             </tr>
             {% endfor %}
