+++ GameMarketplace/templates/base.html
@@ -9,6 +9,9 @@
     <link rel="stylesheet" type="text/css"
         href="https://cdnjs.cloudflare.com/ajax/libs/semantic-ui/2.4.1/semantic.min.css">
     <link rel="stylesheet" type="text/css" href="{% static 'css/style.css' %}">
+</head>
+
+<body>
     <div class="cart">
         {% with total_items=cart|length %}
         {% if cart|length > 0 %}
@@ -21,9 +24,6 @@
         {% endif %}
         {% endwith %}
     </div>
-</head>
-
-<body>
     {% block content %}{% endblock %}
 </body>
