+++ GameMarketplace/templates/cart/detail.html
@@ -1,59 +1,60 @@
-{% extends "shop/base.html" %}
-{% load stics %}
+{% extends "base.html" %}
+{% load static %}
 
 {% block title %}
-    Карзина товаров
+  Корзина покупок
 {% endblock %}
 
 {% block content %}
-    <h1>корзина товаров</h1>
-    <table class = "cart">
-        <thead>
-            <tr>
-                <th>Картинка</th>
-                <th>Товар</th>
-                <th>Обновить кол-во</th>
-                <th>Удалить</th>
-                <th>Кол-во</th>
-                <th>Общая стоимость</th>
-            </tr>
-        </thead>
-        <tbody>
-            {% for item in cart %}
-                {% width product=item.product %}
-                <tr>
-                    <td>
-                        <a href="{{ product.get_absolute_url }}">
-                            <img src="{% if product.image %}{{ product.image.url }}{% else %}{% static 'img/no_img.png' %}{% endif %}">
-                        </a>
-                    </td>
-                    <td>{{ product.name }}</td>
-                    <td>
-                        <form action="{% url 'cart:cart_add' product.id %}" method="post">
-                            {{ item.update_quantity_form.quantity }}
-                            {{ item.update_quantity_form.update }}
-                            <input type="sumbit" value="Обновить">
-                            {% csrf_token %}
-                        </form>
-                    </td>
-                    <td><a href="{% url 'cart:cart_remove' product.id %}">Удалить</a></td>
-                    <td>
-                        {{ item.quantity }}
-                    </td>
-                    <td class="num">${{ item.price }}</td>
-                    <td class="num">${{ item.total_price }}</td>
-                </tr>
-            {% endwith %}
-        {% endfor %}
-        <tr class="total">
-            <td>Всего</td>
-            <td colspan="4"></td>
-            <td colspan="num">${{ cart.get_total_price }}</td>
-        </tr>
-        </tbody>
-    </table>
-    <p class="text-right">
-        <a href="{% url 'shop:product_list' %}" class="button light">В магазин</a>
-        <a href="#" class="button">Оплатить покупку</a>
-    </p>
+  <h1>Корзина покупок</h1>
+  <table class="cart">
+    <thead>
+      <tr>
+        <th>Картинка</th>
+        <th>Товар</th>
+        <th>Обновить кол-во</th>
+        <th>Удалить</th>
+        <th>Кол-во</th>
+        <th>Цена за шт</th>
+        <th>Общая стоимость</th>
+      </tr>
+    </thead>
+    <tbody>
+      {% for item in cart %}
+        {% with product=item.product %}
+          <tr>
+            <td>
+              <a href="{{ product.get_absolute_url }}">
+                <img src="{% if product.img %}{{ product.img.url }}{% else %}{% static 'img/no_image.png' %}{% endif %}">
+              </a>
+            </td>
+            <td>{{ product.name }}</td>
+            <td>
+              <form action="{% url 'cart:cart_add' product.id %}" method="post">
+                {{ item.update_quantity_form.quantity }}
+                {{ item.update_quantity_form.update }}
+                <input type="submit" value="Обновить">
+                {% csrf_token %}
+              </form>
+            </td>
+            <td><a href="{% url 'cart:cart_remove' product.id %}">Удалить</a></td>
+            <td>
+              {{ item.quantity }}
+            </td>
+            <td class="num">${{ item.price }}</td>
+            <td class="num">${{ item.total_price }}</td>
+          </tr>
+        {% endwith %}
+      {% endfor %}
+      <tr class="total">
+        <td>Всего</td>
+        <td colspan="4"></td>
+        <td class="num">${{ cart.get_total_price }}</td>
+      </tr>
+    </tbody>
+  </table>
+  <p class="text-right">
+    <a href="{% url 'market:catalog' %}" class="button light">В магазин</a>
+    <a href="#" class="button">Оформить заказ</a>
+  </p>
 {% endblock %}
